requirements.txt
Click==7.0
Flask==1.1.1
Flask-BasicAuth==0.2.0
Flask-HTTPAuth==4.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
gunicorn==20.1.0
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.10.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
MarkupSafe==1.1.1
numpy==1.19.5
Pillow==6.2.0
psycopg2==2.8.6
pykakasi==2.0.6
secure==0.2.1
SQLAlchemy==1.3.9
tensorflow==2.2.0
tensorflow-cpu==2.2.0
tensorflow-estimator==2.2.0
Werkzeug==0.16.0

